requirements.txt
africastalking==1.2.3
aiohttp==3.7.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
branca==0.4.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
contextlib2==0.6.0.post1
cycler==0.10.0
descartes
Django==3.1.2
django-geojson==3.1.0
django-leaflet==0.27.1
django-rest-framework==0.1.0
django-tables2==2.3.3
djangorestframework==3.12.1
djangorestframework-gis==0.16
Fiona
folium==0.11.0
geographiclib==1.50
geoip2==4.1.0
geopandas
geopy 
gunicorn==20.0.4
idna==2.10
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
maxminddb==2.0.3
multidict==5.0.0
munch==2.5.0
numpy==1.19.3
pandas==1.1.4
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==1.7.1
pyparsing==2.4.7
pyproj
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.55
requests==2.24.0
schema==0.7.2
Shapely
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
yarl==1.6.2
